Добавил связки для полей Тип и цена и время заезда/убытия
